chore: move forwardInnerEventToHost method, minor fixes
